+++ src/tappit/producer.clj
@@ -1,5 +1,45 @@
-(ns tappit.producer
-  (:require [clojure.core.async :as csp]))
+(ns tappit.producer)
+
+;; protocols are low-level and variadic arguments
+;; don't work well with them.
+;; each ok!-based operation has 4 overloads.
+;; eg (ok! thing "this is the name")
+;;    (ok! thing :name "this is the name")
+;;    (ok! thing "this is the name" :diag "diagnostics message)
+;;    (ok! thing "this is the name" :skip "skip message")
+(defprotocol Producer
+  "Test Anything Protocol _Producer_"
+  (plan-for!   [this n] "Creates the 1..N line")
+  (diag!       [this msg] "Creates a diagnostics line")
+  (bail-out!   [this msg] "Bail out / stop testing early")
+  (done!       [this] "Wrap up the protocol")
+
+  (bailed?     [this] "Did we bail already?")
+  (not-bailed? [this] "Are we still good to go?")
+
+  (ok!
+    [this]
+    [this thing]
+    [this thing msg]
+    [this thing flag value]
+    [this thing msg flag value]
+    "Base TAP operation. Assert that something is OK. Produces 'ok' or 'not ok'.")
+  
+  (isa!
+    [this thing pred]
+    [this thing pred msg]
+    [this thing pred flag value]
+    [this thing pred msg flag value]
+    "Predicate-based assertions.")
+  
+  (=!
+    [this thing1 thing2]
+    [this thing1 thing2 msg]
+    [this thing1 thing2 flag value]
+    [this thing1 thing2 msg flag value]
+    "Asserting that two things have the same value."))
+
+;; ----------------------------------------
 
 (defonce ok :ok)
 (defonce not-ok (not ok))
@@ -17,10 +57,12 @@
    thing
    & {name :name
       todo :todo
+      diag :diag
       skip :skip
       :or {name ""
            todo nil
-           skip nil}}]
+           skip nil
+           diag nil}}]
   (if (-bailed? current)
     current
     (let [new (assoc current
@@ -43,6 +85,9 @@
         (not (nil? todo))
         (->diag new (str "TODO "
                          todo))
+
+        (not (nil? diag))
+        (->diag new (str diag))
 
         true new))))
 
@@ -173,6 +218,38 @@
   [t]
   (swap! t ->cleanup)
   ok)
+
+;; ----------------------------------------
+
+(defn create-atom-tap-producer
+  []
+  (let [a (atom {})]
+    (reify Producer
+      (plan-for! [_ n] (->plan-for! a n))
+      (diag! [_ msg] (->diag! a msg))
+      (bail-out! [_ msg] (->bail-out! a msg))
+      (done! [_] (->cleanup! a))
+
+      (bailed? [_] (->bailed? a))
+      (not-bailed? [_] (->not-bailed? a))
+
+      (ok! [_] (->ok! a true))
+      (ok! [_ thing] (->ok! a thing))
+      (ok! [_ thing msg] (->ok! a thing msg))
+      (ok! [_ thing flag value] (->ok! a thing flag value))
+      (ok! [_ thing msg flag value] (->ok! a thing msg flag value))
+
+      (isa! [_ thing pred] (->isa! a thing pred))
+      (isa! [_ thing pred msg] (->isa! a thing pred msg))
+      (isa! [_ thing pred flag value] (->isa! a thing pred flag value))
+      (isa! [_ thing pred msg flag value] (->isa! a thing pred msg flag value))
+
+      (=! [_ thing1 thing2] (->=! a thing1 thing2))
+      (=! [_ thing1 thing2 msg] (->=! a thing1 thing2 msg))
+      (=! [_ thing1 thing2 flag value] (->=! a thing1 thing2 flag value))
+      (=! [_ thing1 thing2 msg flag value] (->=! a thing1 thing2 msg flag value)))))
+
+;; ----------------------------------------
 
 
 (defmacro with-tap!