+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
 ## Planned
+### Changed
+- protocol based api instead of the leaky defmacro for `with-tap!`
+## [Unreleased]
 ### Added
 - generative testing harness for acceptance testing
-
-## [Unreleased]
+- travis ci
 
 ## [0.9.5] - 2016-09-03
 ### Fixed